+++ src/resources/views/inc/sidebar.blade.php
@@ -10,14 +10,14 @@
           </div>
           <div class="pull-left info">
             <p>{{ Auth::user()->name }}</p>
-            <a href="#"><i class="fa fa-circle text-success"></i> Online</a>
+            <a href="#"><i class="fa fa-circle text-success"></i> {{ trans('bytenet-admin-base::base.online') }}</a>
           </div>
         </div>
 
         <!-- search form (Optional) -->
         <form action="#" method="get" class="sidebar-form">
           <div class="input-group">
-            <input type="text" name="q" class="form-control" placeholder="Search...">
+            <input type="text" name="q" class="form-control" placeholder="{{ trans('bytenet-admin-base::base.search') }}...">
             <span class="input-group-btn">
                 <button type="submit" name="search" id="search-btn" class="btn btn-flat"><i class="fa fa-search"></i>
                 </button>
@@ -28,17 +28,25 @@
 
         <!-- sidebar menu: : style can be found in sidebar.less -->
         <ul class="sidebar-menu">
-          <li class="header">{{ trans('bytenet-admin-base::base.administration') }}</li>
+          <li class="header">{{ mb_strtoupper(trans('bytenet-admin-base::base.administration')) }}</li>
           <!-- ================================================ -->
           <!-- ==== Recommended place for admin menu items ==== -->
           <!-- ================================================ -->
-          <li><a href="{{ url('/' . config('bytenet.admin.base.route_prefix') . '/dashboard') }}"><i class="fa fa-dashboard"></i> <span>{{ trans('bytenet-admin-base::base.dashboard') }}</span></a></li>
+          <li class="active">
+              <a href="{{ url('/' . config('bytenet.admin.base.route_prefix') . '/dashboard') }}">
+                  <i class="fa fa-dashboard"></i>
+                  <span>{{ mb_ucfirst(trans('bytenet-admin-base::base.dashboard')) }}</span>
+              </a>
+          </li>
 
+          <li class="header">{{ mb_strtoupper(trans('bytenet-admin-base::base.header')) }}</li>
 
-
-          <li class="header">HEADER</li>
-          <!-- Optionally, you can add icons to the links -->
-          <li class="active"><a href="#"><i class="fa fa-link"></i> <span>Link</span></a></li>
+          <li>
+              <a href="{{ '/' . config('bytenet.admin.base.route_prefix') . '/contacts' }}">
+                  <i class="fa fa-users"></i>
+                  <span>{{ mb_ucfirst(trans('bytenet-admin-contacts::contacts.contacts')) }} {{-- composer! --}}</span>
+              </a>
+          </li>
           <li><a href="#"><i class="fa fa-link"></i> <span>Another Link</span></a></li>
           <li class="treeview">
             <a href="#"><i class="fa fa-link"></i> <span>Multilevel</span>