+++ src/resources/views/inc/sidebar_control.blade.php
@@ -10,7 +10,7 @@
     <div class="tab-content">
         <!-- Home tab content -->
         <div class="tab-pane active" id="control-sidebar-home-tab">
-            <h3 class="control-sidebar-heading">Recent Activity</h3>
+            <h3 class="control-sidebar-heading">{{ trans('bytenet-admin-base::base.recent_activity') }}</h3>
             <ul class="control-sidebar-menu">
                 <li>
                     <a href="javascript:void(0);">