+++ tests/database.php
@@ -62,7 +62,7 @@ return [
             'charset' => 'utf8',
             'collation' => 'utf8_unicode_ci',
             'prefix' => '',
-            'strict' => false,
+            'strict' => true,
             'engine' => null,
         ],
         
@@ -75,7 +75,8 @@ return [
             'charset'   => 'utf8mb4',
             'collation' => 'utf8mb4_unicode_ci',
             'prefix'    => 'btnt_',
-            'strict'    => false,
+            'strict'    => true,
+            'engine'    => null,
         ],
 
         'pgsql' => [